Update user task histories

+++ db/migrate/20150914040030_create_user_tasks.rb
@@ -5,8 +5,8 @@ class CreateUserTasks < ActiveRecord::Migration
       t.references :user_subject, index: true, foreign_key: {on_delete: :cascade}
       t.references :user, index: true, foreign_key: true
       t.integer :progress, default: 0
-      t.integer :spent_time
-      t.integer :estimated_time
+      t.float :spent_time
+      t.float :estimated_time
       t.integer :redmine_task_id
       t.integer :status, default: 0
