+++ db/migrate/20160524010721_create_user_task_histories.rb
@@ -2,8 +2,8 @@ class CreateUserTaskHistories < ActiveRecord::Migration
   def change
     create_table :user_task_histories do |t|
       t.references :user_task, index: true, foreign_key: true
-      t.integer :spent_time
-      t.integer :estimated_time
+      t.float :spent_time
+      t.float :estimated_time
       t.integer :status, default: 0
       t.timestamps null: false
     end
